chore: clippy cleanup — 256→73 warnings across workspace
Mechanical fixes: strip_prefix, identical branches, wildcard patterns,
file create/truncate, redundant guards, if-let simplifications, format
string inlining (via clippy --fix), is_err(), and let-else rewrite.

Remaining 73: unwrap on Result (43, need individual review), complex
types (14, structural), too many args (6, structural), and 4 minor.

+++ crates/shannon-ui/src/repl/query.rs
@@ -578,7 +578,7 @@ pub fn handle_query(repl: &mut Repl, input: &str) -> Result<()> {
                     updated_at: chrono::Utc::now(),
                     total_input_tokens: repl.state.tokens_used,
                     total_output_tokens: 0,
-                    turn_count: repl.current_turn as usize,
+                    turn_count: repl.current_turn,
                     title: None,
                     parent_session_id: None,
                     branch_point_message_index: None,
